+++ CHANGELOG.md
@@ -33,6 +33,7 @@
  * Adding testTuple methods for easily creating Tuple instances to Testing API (thanks xumingming)
  * Trident now throws an error during construction of a topology when try to select fields that don't exist in a stream (thanks xumingming)
  * Compute the capacity of a bolt based on execute latency and #executed over last 10 minutes and display in UI
+ * Storm UI displays exception instead of blank page when there's an error rendering the page (thanks Frostman)
  * Bug fix: Fix deadlock bug due to variant of dining philosophers problem. Spouts now use an overflow buffer to prevent blocking and guarantee that it can consume the incoming queue of acks/fails.
  * Bug fix: Fix race condition in supervisor that would lead to supervisor continuously crashing due to not finding "stormconf.ser" file for an already killed topology
  * Bug fix: bin/storm script now displays a helpful error message when an invalid command is specified